refact: Consolidate config under src/config
