--- src/api/config/swagger.json
@@ -1,60 +1,0 @@
-{
-  "openapi": "3.0.1",
-  "info": {
-    "title": "Njord Basic API",
-    "description": "This is the documentation for Njord API",
-    "contact": {
-      "name": "Invisible Technologies",
-      "url": "https://inv.tech",
-      "email": "engineering@inv.tech"
-    },
-    "license": {
-      "name": "MIT",
-      "url": "https://opensource.org/licenses/MIT"
-    },
-    "version": "0.0.1"
-  },
-  "host": "http://inv-njord.herokuapp.com",
-  "basePath": "/",
-  "tags": [],
-  "schemes": ["http"],
-  "paths": {
-    "/api/v1/event": {
-      "post": {
-        "description": "Path for creating events to the database",
-        "responses": {
-          "200": {
-            "description": "Event was created on the database",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "items": {}
-                }
-              }
-            }
-          },
-          "401": {
-            "description": "Unauthorized request"
-          }
-        }
-      }
-    },
-    "/api-docs": {
-      "get": {
-        "description": "Path for documentation"
-      }
-    }
-  },
-  "securityDefinitions": {
-    "authentication": {
-      "type": "apiKey",
-      "name": "NJORD_API_TOKEN",
-      "in": "header"
-    }
-  },
-  "externalDocs": {
-    "description": "Find out more about Njord",
-    "url": "https://github.com/invisible-tech/njord/blob/master/README.md"
-  }
-}
