+++ src/api/docs/index.ts
@@ -1,7 +1,7 @@
 import * as glob from 'glob'
 import { merge } from 'lodash/fp'
 
-import * as importedSwaggerConfig from '../config/swagger.json'
+import * as importedSwaggerConfig from '../../config/swagger.json'
 
 const paths = glob
   .sync('./*.doc.json', { cwd: __dirname })
