+++ src/api/middlewares/authentication.ts
@@ -3,7 +3,7 @@
 import { NextFunction, Request, Response } from 'express'
 import { UNAUTHORIZED } from 'http-status'
 
-import { NJORD_API_TOKEN } from '../config'
+import { NJORD_API_TOKEN } from '../../config'
 
 export const authentication = (req: Request, res: Response, next: NextFunction) => {
   if (req.headers.authorization === `Bearer ${NJORD_API_TOKEN}`) next()
