+++ src/config/index.ts
@@ -1,9 +1,7 @@
-'use strict'
+import * as env from 'env-var'
 
-export const Config = {
-  NODE_ENV: process.env.NODE_ENV,
-  PORT: process.env.PORT || 3006,
-  HEIMDALL_URL: process.env.HEIMDALL_URL,
-  DB_CONNECTION_STRING: process.env.DB_CONNECTION_STRING,
-  DB_TEST_CONNECTION_STRING: process.env.DB_TEST_CONNECTION_STRING,
-}
+export const NODE_ENV = env.get('NODE_ENV', 'developer').asString()
+export const PORT: number = env.get('PORT', '3006').asInt()
+export const NJORD_API_TOKEN: string = env.get('NJORD_API_TOKEN').required().asString()
+export const DB_CONNECTION_STRING: string = env.get('DB_CONNECTION_STRING').required().asString()
+export const DB_TEST_CONNECTION_STRING: string = env.get('DB_TEST_CONNECTION_STRING').required().asString()
